small file_uploader refactor
