+++ conans/client/rest/rest_client_v1.py
@@ -7,8 +7,9 @@ from six.moves.urllib.parse import parse_qs, urljoin, urlparse, urlsplit
 from conans.client.remote_manager import check_compressed_files
 from conans.client.rest.client_routes import ClientV1Router
 from conans.client.rest.download_cache import CachedFileDownloader
+from conans.client.rest.file_uploader import FileUploader
 from conans.client.rest.rest_client_common import RestCommonMethods, handle_return_deserializer
-from conans.client.rest.uploader_downloader import FileDownloader, FileUploader
+from conans.client.rest.uploader_downloader import FileDownloader
 from conans.errors import ConanException, NotFoundException, NoRestV2Available, \
     PackageNotFoundException
 from conans.model.info import ConanInfo
